+++ enterprise/frontend/src/metabase-enterprise/data-studio/common/components/PaneHeader/PaneHeader.module.css
@@ -1,4 +1,4 @@
-.ModeSwitcher {
+.ProfileLink {
   @media screen and (max-width: calc(50rem + 300px)) {
     position: fixed;
     right: 3.5rem;
