+++ enterprise/frontend/src/metabase-enterprise/dependencies/pages/DependencyGraphPage/DependencyGraphPage.module.css
@@ -1,4 +1,4 @@
-.ModeSwitcher {
+.ProfileLink {
   position: absolute;
   top: 1.25rem;
   right: 3.5rem;
