+++ frontend/src/metabase/nav/components/ProfileLink/ProfileLink.unit.spec.tsx
@@ -1,13 +1,15 @@
 import userEvent from "@testing-library/user-event";
 import dayjs from "dayjs";
+import { Route } from "react-router";
 
+import { setupEnterprisePlugins } from "__support__/enterprise";
 import { setupBugReportingDetailsEndpoint } from "__support__/server-mocks";
 import { mockSettings } from "__support__/settings";
-import { renderWithProviders, screen, waitFor } from "__support__/ui";
-import { ProfileLink } from "metabase/nav/components/ProfileLink";
+import { renderWithProviders, screen, waitFor, within } from "__support__/ui";
 import type { HelpLinkSetting } from "metabase-types/api";
 import {
   createMockMetabaseInfo,
+  createMockTokenFeatures,
   createMockTokenStatus,
   createMockUser,
 } from "metabase-types/api/mocks";
@@ -16,14 +18,20 @@ import {
   createMockAdminState,
 } from "metabase-types/store/mocks";
 
+import { ProfileLink } from "./ProfileLink";
+
+const USER = createMockUser();
+
 const REGULAR_ITEMS = [
-  "Account settings",
+  USER.first_name as string,
+  USER.email,
   "Help",
-  "About Metabase",
   "Sign out",
 ];
-const ADMIN_ITEMS = [...REGULAR_ITEMS, "Admin settings"];
+const ADMIN_ITEMS = [...REGULAR_ITEMS, "Main app", "Admin"];
 const HOSTED_ITEMS = [...ADMIN_ITEMS];
+
+const WITH_DATA_STUDIO = [...ADMIN_ITEMS, "Data studio"];
 
 const adminNavItem = {
   name: `People`,
@@ -31,28 +39,41 @@ const adminNavItem = {
   key: "people",
 } as const;
 
-function setup({
+async function setup({
   isAdmin = false,
   isHosted = false,
   isPaidPlan = true,
+  hasDataStudio = false,
   helpLinkSetting = "metabase",
   helpLinkCustomDestinationSetting = "https://custom-destination.com/help",
   instanceCreationDate = dayjs().toISOString(),
+  initialRoute = "/",
 }: {
   isAdmin?: boolean;
   isHosted?: boolean;
   isPaidPlan?: boolean;
+  hasDataStudio?: boolean;
   helpLinkSetting?: HelpLinkSetting;
   helpLinkCustomDestinationSetting?: string;
   instanceCreationDate?: string;
-}) {
+  initialRoute?: string;
+} = {}) {
+  setupBugReportingDetailsEndpoint();
+
   const settings = mockSettings({
     "is-hosted?": isHosted,
     "token-status": createMockTokenStatus({ valid: isPaidPlan }),
     "help-link": helpLinkSetting,
     "help-link-custom-destination": helpLinkCustomDestinationSetting,
     "instance-creation": instanceCreationDate,
+    "token-features": createMockTokenFeatures({
+      data_studio: hasDataStudio,
+    }),
   });
+
+  if (hasDataStudio) {
+    setupEnterprisePlugins();
+  }
 
   const admin = createMockAdminState({
     app: createMockAdminAppState({
@@ -60,29 +81,54 @@ function setup({
     }),
   });
 
-  return renderWithProviders(<ProfileLink />, {
-    storeInitialState: {
-      admin,
-      settings,
-      currentUser: createMockUser({ is_superuser: isAdmin }),
+  renderWithProviders(
+    <>
+      <Route path="/" component={ProfileLink} />
+      <Route path="/admin" component={ProfileLink} />
+      <Route path="/data-studio" component={ProfileLink} />
+    </>,
+    {
+      withRouter: true,
+      initialRoute,
+      storeInitialState: {
+        admin,
+        settings,
+        currentUser: { ...USER, is_superuser: isAdmin },
+      },
     },
-  });
+  );
+
+  await openProfileLink();
 }
 
-function setupHosted(opts = {}) {
+async function setupHosted(opts = {}) {
   return setup({ ...opts, isHosted: true });
 }
 
 describe("ProfileLink", () => {
-  beforeEach(() => {
-    setupBugReportingDetailsEndpoint();
+  it("should render standard links", async () => {
+    await setup();
+
+    // Should always render a profile link
+    expect(
+      await screen.findByText(USER.first_name as string),
+    ).toBeInTheDocument();
+    expect(await screen.findByText(USER.email)).toBeInTheDocument();
+
+    //Should render a help submenu
+    expect(
+      await screen.findByRole("menuitem", { name: "Help" }),
+    ).toBeInTheDocument();
+
+    // Should render logout
+    expect(
+      await screen.findByRole("menuitem", { name: "Sign out" }),
+    ).toBeInTheDocument();
   });
 
   describe("self-hosted", () => {
     it("should show the proper set of items for normal users", async () => {
-      setup({ isAdmin: false });
-
-      await openMenu();
+      await setup({ isAdmin: false });
 
       REGULAR_ITEMS.forEach((title) => {
         expect(screen.getByText(title)).toBeInTheDocument();
@@ -91,9 +137,7 @@ describe("ProfileLink", () => {
     });
 
     it("should show the proper set of items for admin users", async () => {
-      setup({ isAdmin: true });
-
-      await openMenu();
+      await setup({ isAdmin: true });
 
       ADMIN_ITEMS.forEach((title) => {
         expect(screen.getByText(title)).toBeInTheDocument();
@@ -101,11 +145,35 @@ describe("ProfileLink", () => {
     });
   });
 
+  describe("current app", () => {
+    it("should update it's apps section as you navigate", async () => {
+      await setup({ isAdmin: true, hasDataStudio: true });
+
+      await assertActiveApp("main");
+
+      await userEvent.click(await getAdminMenuItem());
+      await openProfileLink();
+      await waitFor(() => assertActiveApp("admin"));
+
+      await userEvent.click(await getDataStudioMenuItem());
+      await openProfileLink();
+      await assertActiveApp("data-studio");
+
+      await userEvent.click(await getMainAppMenuItem());
+      await openProfileLink();
+      await assertActiveApp("main");
+    });
+  });
+
+  it("should correctly hightlight admin app", async () => {
+    await setup({ isAdmin: true, hasDataStudio: true, initialRoute: "/admin" });
+
+    await waitFor(() => assertActiveApp("admin"));
+  });
+
   describe("hosted", () => {
     it("should show the proper set of items for normal users", async () => {
-      setupHosted({ isAdmin: false });
-
-      await openMenu();
+      await setupHosted({ isAdmin: false });
 
       REGULAR_ITEMS.forEach((title) => {
         expect(screen.getByText(title)).toBeInTheDocument();
@@ -114,9 +182,7 @@ describe("ProfileLink", () => {
     });
 
     it("should show the proper set of items for admin users", async () => {
-      setupHosted({ isAdmin: true });
-
-      await openMenu();
+      await setupHosted({ isAdmin: true });
 
       HOSTED_ITEMS.forEach((title) => {
         expect(screen.getByText(title)).toBeInTheDocument();
@@ -124,99 +190,148 @@ describe("ProfileLink", () => {
     });
   });
 
-  describe("'How to use Metabase' link", () => {
-    it("should render if the instance was created more than 30 days ago", async () => {
-      setup({
-        instanceCreationDate: dayjs().subtract(42, "days").toISOString(),
+  describe("with data studio", () => {
+    it("should show data studio app when apropriate", async () => {
+      await setup({ isAdmin: true, hasDataStudio: true });
+
+      WITH_DATA_STUDIO.forEach((title) => {
+        expect(screen.getByText(title)).toBeInTheDocument();
       });
-
-      await openMenu();
-      expect(screen.getByText("How to use Metabase")).toBeInTheDocument();
-    });
-
-    it("should not render for new instances (younger than 30 days)", async () => {
-      setup({
-        instanceCreationDate: dayjs().subtract(14, "days").toISOString(),
-      });
-
-      await openMenu();
-      expect(screen.queryByText("How to use Metabase")).not.toBeInTheDocument();
     });
   });
 
-  describe("help link", () => {
-    describe("when the setting is set to hidden", () => {
-      it("should return not be visible", async () => {
-        setup({
-          helpLinkSetting: "hidden",
+  describe("help submenu", () => {
+    describe("'How to use Metabase' link", () => {
+      it("should render if the instance was created more than 30 days ago", async () => {
+        await setup({
+          instanceCreationDate: dayjs().subtract(42, "days").toISOString(),
         });
-        await openMenu();
+        await openHelpSubmenu();
 
-        const link = screen.queryByRole("menuitem", { name: /help/i });
+        expect(screen.getByText("How to use Metabase")).toBeInTheDocument();
+      });
 
-        expect(link).not.toBeInTheDocument();
+      it("should not render for new instances (younger than 30 days)", async () => {
+        await setup({
+          instanceCreationDate: dayjs().subtract(14, "days").toISOString(),
+        });
+        await openHelpSubmenu();
+
+        expect(
+          screen.queryByText("How to use Metabase"),
+        ).not.toBeInTheDocument();
       });
     });
 
-    describe("when the setting is `custom`", () => {
-      it("should return  the custom destination", async () => {
-        setup({
-          helpLinkSetting: "custom",
-          helpLinkCustomDestinationSetting: "https://custom.example.org/help",
-        });
-        await openMenu();
-
-        const link = screen.getByRole("menuitem", { name: /help/i });
-
-        expect(link).toBeInTheDocument();
-        expect(link).toHaveProperty("href", "https://custom.example.org/help");
-      });
-    });
-
-    describe("when the setting is `metabase`", () => {
-      describe("when admin on paid plan", () => {
-        it("should return the default /help-premium link", async () => {
-          setup({
-            isAdmin: true,
-            isPaidPlan: true,
-            helpLinkSetting: "metabase",
+    describe("help link", () => {
+      describe("when the setting is set to hidden", () => {
+        it("should return not be visible", async () => {
+          await setup({
+            helpLinkSetting: "hidden",
           });
-          await openMenu();
-          const link = screen.getByRole("menuitem", { name: /help/i });
+          await openHelpSubmenu();
 
-          expect(link).toBeInTheDocument();
-          const mockBugReportDetails = encodeURIComponent(
-            JSON.stringify(createMockMetabaseInfo()),
-          );
-          const expected = `https://www.metabase.com/help-premium?utm_source=in-product&utm_medium=menu&utm_campaign=help&instance_version=v1&diag=${mockBugReportDetails}`;
-          await waitFor(() => expect(link).toHaveProperty("href", expected));
+          const link = screen.queryByRole("menuitem", { name: /get help/i });
 
-          expect(link).toHaveProperty("href", expected);
+          expect(link).not.toBeInTheDocument();
         });
       });
 
-      describe("when non admin", () => {
-        it("should return the default /help link", async () => {
-          setup({
-            isAdmin: false,
-            isPaidPlan: true,
-            helpLinkSetting: "metabase",
+      describe("when the setting is `custom`", () => {
+        it("should return the custom destination", async () => {
+          await setup({
+            helpLinkSetting: "custom",
+            helpLinkCustomDestinationSetting: "https://custom.example.org/help",
           });
-          await openMenu();
-          const link = screen.getByRole("menuitem", { name: /help/i });
+          await openHelpSubmenu();
+
+          const link = screen.getByRole("menuitem", { name: /get help/i });
 
           expect(link).toBeInTheDocument();
           expect(link).toHaveProperty(
             "href",
-            "https://www.metabase.com/help?utm_source=in-product&utm_medium=menu&utm_campaign=help&instance_version=v1",
+            "https://custom.example.org/help",
           );
+        });
+      });
+
+      describe("when the setting is `metabase`", () => {
+        describe("when admin on paid plan", () => {
+          it("should return the default /help-premium link", async () => {
+            await setup({
+              isAdmin: true,
+              isPaidPlan: true,
+              helpLinkSetting: "metabase",
+            });
+
+            await openHelpSubmenu();
+
+            const link = screen.getByRole("menuitem", { name: /get help/i });
+
+            expect(link).toBeInTheDocument();
+            const mockBugReportDetails = encodeURIComponent(
+              JSON.stringify(createMockMetabaseInfo()),
+            );
+            const expected = `https://www.metabase.com/help-premium?utm_source=in-product&utm_medium=menu&utm_campaign=help&instance_version=v1&diag=${mockBugReportDetails}`;
+            await waitFor(() => expect(link).toHaveProperty("href", expected));
+
+            expect(link).toHaveProperty("href", expected);
+          });
+        });
+
+        describe("when non admin", () => {
+          it("should return the default /help link", async () => {
+            await setup({
+              isAdmin: false,
+              isPaidPlan: true,
+              helpLinkSetting: "metabase",
+            });
+
+            await openHelpSubmenu();
+
+            const link = screen.getByRole("menuitem", { name: /get help/i });
+
+            expect(link).toBeInTheDocument();
+            expect(link).toHaveProperty(
+              "href",
+              "https://www.metabase.com/help?utm_source=in-product&utm_medium=menu&utm_campaign=help&instance_version=v1",
+            );
+          });
         });
       });
     });
   });
 });
 
-const openMenu = async () => {
-  await userEvent.click(screen.getByRole("img", { name: /gear/i }));
+const openProfileLink = async () => {
+  await userEvent.click(screen.getByRole("img", { name: /mode/i }));
   await screen.findByRole("menu");
 };
+
+const openHelpSubmenu = async () =>
+  await userEvent.click(await screen.findByRole("menuitem", { name: "Help" }));
+
+const assertActiveApp = async (current: "main" | "admin" | "data-studio") => {
+  expect(
+    await within(await getMainAppMenuItem()).findByRole("img", {
+      name: current === "main" ? /check_filled/i : /dashboard/i,
+    }),
+  ).toBeInTheDocument();
+  expect(
+    await within(await getAdminMenuItem()).findByRole("img", {
+      name: current === "admin" ? /check_filled/i : /io/i,
+    }),
+  ).toBeInTheDocument();
+  expect(
+    await within(await getDataStudioMenuItem()).findByRole("img", {
+      name: current === "data-studio" ? /check_filled/i : /table/i,
+    }),
+  ).toBeInTheDocument();
+};
+
+const getMainAppMenuItem = () =>
+  screen.findByRole("menuitem", { name: /main app/i });
+const getAdminMenuItem = () =>
+  screen.findByRole("menuitem", { name: /admin/i });
+const getDataStudioMenuItem = () =>
+  screen.findByRole("menuitem", { name: /data studio/i });
